feat: refactor folder structure, add setAnimationFrameTimeout
