+++ src/useSwitchTransition/index.ts
@@ -1,0 +1,1 @@
+export function useSwitchTransition() {}
